Semantic Embedding (unfinished 9.12.22)
This project was produced for an assignment, in which I was required to build a prototype question answering system using the semantic information embedded in a Distributional Semantic Model generated by the Global Vectors for Word Representation method. The system was intended to find a list of words semantically close to the word provided.