Minigame Project (25.11.22)
This project produces the controls of a game, in which one is either playing as a Basilisk or a Goblin. Using the methods within abstract superclass 'Creature', the Basilisk and Goblin different methods in order to attack and move. The method 'isInRange' is abstract and differs for the Basilisk vs the Goblin, who have different ranges of attack.