Simple Battleship Command Line Game
This battleships game is produced to receive user input through command line. The user inputs an integer between 1 and 6, and the program calculates whether this hits or misses the 'battleship'. When all 3 parts of the battleship are hit, the user wins. The application tallies the number of attempts this took the user which is taken as their score.